requirements.txt
cassandra_driver==3.29.2
cassio==0.1.10
langchain_core==0.3.15
python-dotenv==1.0.1
Requests==2.32.3
unstructured==0.16.0
unstructured==0.16.4
